requirements.txt
Fastapi
openai
requests
pydantic
Jinja2
markdown2
uvicorn
python-dotenv
python-multipart